requirements.txt
# Core FastAPI
fastapi==0.104.1
starlette==0.27.0
uvicorn==0.24.0
click==8.1.7
typing-extensions==4.8.0
annotated-types==0.6.0
h11==0.14.0
httptools==0.6.1
watchfiles==0.21.0
websockets==12.0
anyio==3.7.1
sniffio==1.3.0
exceptiongroup==1.1.3
idna==3.4
certifi==2023.7.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic-core==2.14.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
greenlet==3.0.1
alembic==1.12.1

# Cache & Queue
redis==5.0.1
async-timeout==4.0.3
celery==5.3.4
kombu==5.3.4
billiard==4.2.0
vine==5.1.0

# Auth & External Services
firebase-admin==6.2.0
httpx==0.24.1
httpcore==0.17.3
supabase==1.2.0
google-cloud-firestore==2.13.1
google-auth==2.23.4
cachetools==5.3.2
rsa==4.9
pyasn1==0.5.0
pyasn1-modules==0.3.0

# Search & Utils
meilisearch==0.28.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
loguru==0.7.2
boto3==1.34.0
python-dotenv==1.0.0
botocore==1.34.0
jmespath==1.0.1
s3transfer==0.10.0
urllib3==2.0.7
requests==2.31.0
charset-normalizer==3.3.2
cryptography==41.0.7
ecdsa==0.18.0
pytz==2023.3
six==1.16.0
markdown==3.5.1
mako==1.3.0
markupsafe==2.1.3